Add a way to destroy NetworkController in tests
During NetworkController's initialization process, async operations are
run out of band:

1. We make a request to Ethereum to check Infura availability, but never
  wait until this request is finished
2. The PollingBlockTracker is started, but we never stop it

This makes it difficult to mock HTTP requests in tests for
NetworkController, as requests may be made after tests end.

This commit adds a hook to NetworkController that other methods can use
to know when NetworkController is fully initialized with a provider, and
also adds a `destroy` method which waits for the provider to be
available before stopping the block tracker.